Rename file
This file will contain all auth functions. Renamed and moved
appropriately.
